+++ client/common/components/EventList/ProjectColumn/ResolveProjectModal/types.tsx
@@ -1,0 +1,7 @@
+import { IProject, ITimeEntry } from 'interfaces';
+export interface IResolveProjectModalProps {
+    event: ITimeEntry;
+    isOpen?: boolean;
+    onDismiss?: () => void;
+    onProjectSelected?: (project: IProject) => void;
+}