+++ client/common/components/List/ListHeader/index.tsx
@@ -5,10 +5,15 @@ import { IRenderFunction } from 'office-ui-fabric-react/lib/Utilities';
 import * as React from 'react';
 import { IListProps } from '../IListProps';
 
-export const ListHeader = (headerProps: IDetailsHeaderProps, defaultRender: IRenderFunction<IDetailsHeaderProps>, props: IListProps, onSearch: (event: any, term: string) => void) => {
+export const ListHeader = (headerProps: IDetailsHeaderProps, defaultRender: IRenderFunction<IDetailsHeaderProps>, props: IListProps, onSearch: (searchTerm: string) => void) => {
     return (
         <Sticky stickyPosition={StickyPositionType.Header} isScrollSynced={true}>
-            {props.searchBox && <SearchBox styles={{ field: { fontSize: '10pt', letterSpacing: '1px' }, root: { maxWidth: 320 } }} {...props.searchBox} onChange={onSearch} />}
+            {props.searchBox && (
+                <SearchBox
+                    {...props.searchBox}
+                    styles={{ field: { fontSize: '10pt', letterSpacing: '1px' }, root: { maxWidth: 320 } }}
+                    onChange={(_, newValue) => onSearch(newValue)} />)
+            }
             {defaultRender(headerProps)}
         </Sticky>
     );