+++ client/common/components/SearchProject/types.ts
@@ -1,0 +1,7 @@
+import { ICustomer } from 'interfaces/ICustomer';
+
+export interface ISearchProjectProps {
+    onSelected: any;
+    customer: ICustomer;
+    placeholder: string;
+}