+++ client/common/components/index.ts
@@ -3,3 +3,5 @@ export * from './IconPicker';
 export * from './List';
 export * from './ScrollablePaneWrapper';
 export * from './UserMessage';
+export * from './SearchProject';
+export * from './EventList';