+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for split weeks #232
 - Added user column to project overview #224
 - Consolidating confirm buttons #270
+- Confirm hours disabled when there's unmatched events #268
 
 ### Fixed
 - Sorted customer options alphabetically by name in summary view(s) #260